pyproject.toml
[project]
name = "django-action"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "qe-present", email = "2664481691@qq.com" }
]
dependencies = [
    "djangorestframework>=3.16.1",
    "gunicorn>=23.0.0",
    "uvicorn>=0.35.0",
    "dotenv>=0.9.9",
    "celery>=5.5.3",
    "redis>=6.4.0",
    "regex>=2025.9.18",
    "django-cors-headers>=4.9.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/django_action"]
